requirements.txt
# Core LLM and Embeddings
ollama>=0.1.0

# Vector Database
chromadb>=0.4.0

# Knowledge Graph
networkx>=3.0
pyvis>=0.3.2

# Keyword Search
rank-bm25>=0.2.2

# Document Processing
pypdf2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
markdown>=3.5.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# UI Framework
streamlit>=1.30.0
plotly>=5.18.0
streamlit-agraph>=0.0.45
pyvis>=0.3.2

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
